Partial algorithm rewrite (not fully tested)
This cuts down on the amount of text iterated. Text is iterated backwards until the start of the block, then text measuring and tabstop stretching is began from that position.

Previously it iterated over the entire block in 3 chunks (before, start to end, and after) to determine block boundaries and compute the max number of tabs used...then reiterated the entire block to measure text width.